+++ src/main/java/lotto/WebUILottoApplication.java
@@ -7,9 +7,9 @@ import lotto.domain.lottogenerator.LottoGenerator;
 import lotto.domain.lottogenerator.ManualLottoGeneratingStrategy;
 import lotto.domain.paymentinfo.CountOfLotto;
 import lotto.domain.paymentinfo.Payment;
-import lotto.domain.service.LottoResultService;
-import lotto.domain.service.LottoService;
-import lotto.domain.service.PaymentInfoService;
+import lotto.service.LottoResultService;
+import lotto.service.LottoService;
+import lotto.service.PaymentInfoService;
 import lotto.exception.IllegalParameterException;
 import spark.ModelAndView;
 import spark.template.handlebars.HandlebarsTemplateEngine;
