+++ src/test/java/lotto/dao/JdbcConnectorTest.java
@@ -1,6 +1,6 @@
-package lotto.domain.dao;
+package lotto.dao;
 
-import lotto.domain.dao.utils.JdbcConnector;
+import lotto.dao.utils.JdbcConnector;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
