+++ src/test/java/lotto/dao/PaymentInfoTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.dao;
+package lotto.dao;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
